Conflicts:
	examples/CMakeLists.txt

+++ source/dtEntityQtWidgets/spawnerstore.cpp
@@ -20,8 +20,8 @@
 
 #include <dtEntityQtWidgets/spawnerstore.h>
 #include <dtEntity/applicationcomponent.h>
-#include <dtEntity/basemessages.h>
 #include <dtEntity/cameracomponent.h>
+#include <dtEntity/commandmessages.h>
 #include <dtEntity/entity.h>
 #include <dtEntity/entitymanager.h>
 #include <dtEntity/layerattachpointcomponent.h>
@@ -497,7 +497,7 @@ namespace dtEntityQtWidgets
       std::vector<dtEntity::EntityId> ids;
       mtsystem->GetSpawnerCreatedEntities(name.toStdString(), ids);
 
-      for(unsigned int i = 0; i < ids.size(); ++i)
+      for(size_t i = 0; i < ids.size(); ++i)
       {
          dtEntity::RequestEntitySelectMessage msg;
          msg.SetAboutEntityId(ids[i]);
